+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Readme - Add a CI build badge in the readme.
 - Release Note - Publish well formated release note depending on current path.
 - Release Script - Centralize the script to release a project.
+- Technical - Add a coverage badge.